refactor base cdc storage
